separate createElement function into module and add comments
