+++ index.html
@@ -19,12 +19,6 @@
       <img src="./images/icon-dbl-left.svg" alt="close icon" id="sidebar-close-btn" class="open-close-btn" />
       <div class="content">Your sidebar will go in this div</div>
       
-      <div id="select-countdown-sidebar" class="content">
-        <h1><strong>SELECT</strong> COUNTDOWN</h1>
-        
-        <div id="countdown-list">
-  
-        </div>
       </div>
 
     </aside>
@@ -54,7 +48,5 @@
   </footer>
 
   <script src="./resources/js/closeSidebar.js"></script>
-  <script src="./resources/js/createCountdownItem.js"></script>
-
 </body>
 </html>
