+++ resources/html/sideBars/createCountdownBar.html
@@ -2,6 +2,7 @@
   <h1><strong>SELECT</strong> COUNTDOWN</h1>
   
   <div id="countdown-list">
-
+    <!--This list will be populated by the createCountdownItem.js file, where countdown cards are generated from the countdown objects (will eventually be from data from the server)-->
+    <script src="./resources/js/createCountdownItem.js" type="module"></script>
   </div>
 </div>
