+++ styles.css
@@ -172,19 +172,24 @@ aside .content {
 }
 
 .name {
+  align-self: center;
+  width: 200%;
+  font-size: 1.4rem;
+  font-weight: bold;
   grid-row: 0 / span 2;
 }
-
 .date {
+  align-self: center;
   grid-column: 1 / span 1;
   grid-row: 2 / span 1;
 }
-
 .time {
+  align-self: center;
   grid-column: 2 / span 1;
   grid-row: 2 / span 1;
 }
 
+/*styles for the "Add New" card*/
 .item.add-new {
   display: flex;
   border: 2px red dashed;
